requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==36.0.0
h11==0.12.0
idna==3.3
mypy-extensions==0.4.3
numpy==1.21.4
outcome==1.1.0
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.0
urllib3==1.26.7
webdriver-manager==3.5.2
wsproto==1.0.0
